Simple proof of concept web-app
Using Flask to interface with python library.  This takes a hard-coded
url and passes it to our TagParser to get the tags.  We render the html
text into a textarea and then use a jquery plugin to highlight the tags
in the textarea.  The highlighting of the text is made possible by
using the indexs of start/end kept in our Tag objects that the
TagParser aggregates.